requirements.txt
langchain
langchain_community
langchain_huggingface
faiss-cpu
pypdf
huggingface_hub
flask   
python-dotenv
sentence-transformers 
groq
langchain-groq
jq